# swagger_ui/openapi_ms1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Investment Data Service API",
    "description": "Documentação completa das rotas do microserviço investment_data_service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor local"
    }
  ],
  "paths": {
...
    "/clientes": {
      "post": {
        "summary": "Criar cliente",
        "tags": ["Clientes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string" },
                  "cpf": { "type": "string" },
                  "email": { "type": "string" },
                  "telefone": { "type": "string" },
                  "correntista": { "type": "integer" },
                  "score_credito": { "type": "number" },
                  "saldo_cc": { "type": "number" },
                  "admin": { "type": "integer" },
                  "senha": { "type": "string" }
                },
                "required": ["cpf", "email", "telefone", "senha"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado"
          },
          "400": {
            "description": "Dados duplicados (CPF, email, telefone)"
          }
        }
      }
    },
    "/clientes/transacao": {
      "post": {
        "summary": "Criar transação para cliente",
        "tags": ["Clientes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliente_id": { "type": "string" },
                  "tipo": { 
                    "type": "string",
                    "enum": ["deposito", "saque", "transferencia", "investimento", "ajuste"]
                  },
                  "valor": { "type": "number" },
                  "descricao": { "type": "string" }
                },
                "required": ["cliente_id", "tipo", "valor"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transação criada"
          },
          "400": {
            "description": "Cliente não existe ou dados inválidos"
          }
        }
      }
    },
    "/clientes/buscar/cpf/{cpf}": {
      "get": {
        "summary": "Buscar cliente pelo CPF",
        "tags": ["Clientes"],
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
    "/clientes/buscar/telefone/{telefone}": {
      "get": {
        "summary": "Buscar cliente pelo telefone",
        "tags": ["Clientes"],
        "parameters": [
          {
            "name": "telefone",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
    "/clientes/buscar/email/{email}": {
      "get": {
        "summary": "Buscar cliente pelo e-mail",
        "tags": ["Clientes"],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente encontrado"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
    "/clientes/cpf/atualiza/{cpf}": {
      "patch": {
        "summary": "Atualizar cliente pelo CPF",
        "tags": ["Clientes"],
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string" },
                  "cpf": { "type": "string" },
                  "email": { "type": "string" },
                  "telefone": { "type": "string" },
                  "correntista": { "type": "integer" },
                  "score_credito": { "type": "number" },
                  "saldo_cc": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cliente atualizado com sucesso"
          },
          "400": {
            "description": "CPF, email ou telefone já cadastrados"
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
    "/clientes/delete/cpf/{cpf}": {
      "delete": {
        "summary": "Deletar cliente pelo CPF",
        "tags": ["Clientes"],
        "parameters": [
          {
            "name": "cpf",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sucesso": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cliente não encontrado"
          }
        }
      }
    },
    "/clientes/login": {
      "post": {
        "summary": "Login de cliente",
        "tags": ["Clientes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cpf": { "type": "string" },
                  "senha": { "type": "string" }
                },
                "required": ["cpf", "senha"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login efetuado"
          },
          "400": {
            "description": "Campos faltando"
          },
          "401": {
            "description": "CPF ou senha incorretos"
          }
        }
      }
    },
    "/clientes/alterar-senha": {
      "patch": {
        "summary": "Alterar senha do cliente",
        "tags": ["Clientes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cpf": { "type": "string" },
                  "senha_atual": { "type": "string" },
                  "nova_senha": { "type": "string" }
                },
                "required": ["cpf", "senha_atual", "nova_senha"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso"
          },
          "400": {
            "description": "Campos obrigatórios não enviados"
          },
          "401": {
            "description": "Senha atual incorreta"
          }
        }
      }
    },
...
    "/api/acoes": {
      "post": {
        "summary": "Criar investimento de ação",
        "tags": ["API Ações"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliente_id": { "type": "string" },
                  "ticker": { "type": "string" },
                  "quantidade": { "type": "number" },
                  "preco_unitario": { "type": "number" },
                  "custo_total": { "type": "number" }
                },
                "required": ["cliente_id"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Investimento criado"
          },
          "400": {
            "description": "cliente_id ausente"
          }
        }
      }
    },
    "/api/acoes/cliente/{cliente_id}": {
      "get": {
        "summary": "Listar ações de um cliente",
        "tags": ["API Ações"],
        "parameters": [
          {
            "name": "cliente_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ações do cliente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id_investimento": { "type": "integer" },
                      "cliente_id": { "type": "string" },
                      "ticker": { "type": "string" },
                      "quantidade": { "type": "number" },
                      "preco_unitario": { "type": "number" },
                      "custo_total": { "type": "number" },
                      "data": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/acoes/{id_investimento}": {
      "get": {
        "summary": "Buscar ação por ID",
        "tags": ["API Ações"],
        "parameters": [
          {
            "name": "id_investimento",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Ação encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_investimento": { "type": "integer" },
                    "cliente_id": { "type": "string" },
                    "ticker": { "type": "string" },
                    "quantidade": { "type": "number" },
                    "preco_unitario": { "type": "number" },
                    "custo_total": { "type": "number" },
                    "data": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Investimento não encontrado"
          }
        }
      },

      "patch": {
        "summary": "Atualizar investimento de ação",
        "tags": ["API Ações"],
        "parameters": [
          {
            "name": "id_investimento",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantidade": { "type": "number" },
                  "custo_total": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Investimento atualizado"
          }
        }
      },

      "delete": {
        "summary": "Deletar investimento de ação",
        "tags": ["API Ações"],
        "parameters": [
          {
            "name": "id_investimento",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Investimento excluído com sucesso"
          }
        }
      }
    },
    "/api/acoes/transacao": {
      "post": {
        "summary": "Registrar operação de compra/venda",
        "tags": ["API Ações"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliente_id": { "type": "string" },
                  "ticker": { "type": "string" },
                  "tipo": { 
                    "type": "string",
                    "enum": ["compra", "venda"]
                  },
                  "quantidade": { "type": "number" },
                  "preco_unitario": { "type": "number" },
                  "custo_total": { "type": "number" }
                },
                "required": ["cliente_id", "ticker", "tipo", "quantidade", "preco_unitario"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação registrada"
          }
        }
      }
    }
  },  
...
}